refactor: separate ChatInput
